refactor: Update Relevance component heading and add new dependencies
This commit updates the Relevance component in Relevance.jsx to change the heading from "Relevance Chart" to "Relevance". This change provides a more concise and descriptive heading for the component.

Additionally, the commit adds new dependencies in package.json. The new dependencies include:
- d3 v7.9.0
- react-vega v7.6.0
- vega v5.29.0
- vega-embed v6.25.0
- vega-lite v5.18.1

These dependencies are necessary for implementing data visualization features in the application, such as charts and graphs.